+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New features
+
+* [#70](https://github.com/tboox/tbox/issues/70): Add `tb_stream_init_from_sock_ref()` to open a given socket as stream
+
 ## v1.6.3
 
 ### New features
@@ -164,6 +168,10 @@
 # 更新日志
 
 ## master (开发中)
+
+### 新特性
+
+* [#70](https://github.com/tboox/tbox/issues/70): 添加`tb_stream_init_from_sock_ref()`接口去直接打开一个socket作为stream去读取数据。
 
 ## v1.6.3